Took more functions out of specfun.h
