+++ scipy/special/special/airy.h
@@ -114,6 +114,313 @@ namespace detail {
 
 } // namespace detail
 
+inline void airyb(double x, double *ai, double *bi, double *ad, double *bd) {
+
+    // =======================================================
+    // Purpose: Compute Airy functions and their derivatives
+    // Input:   x  --- Argument of Airy function
+    // Output:  AI --- Ai(x)
+    //          BI --- Bi(x)
+    //          AD --- Ai'(x)
+    //          BD --- Bi'(x)
+    // =======================================================
+
+    int k, km, km2, kmax;
+    double ck[52], dk[52];
+    double xa, xq, xm, fx, r, gx, df, dg, sai, sad, sbi, sbd, xp1, xr2;
+    double xf, rp, xar, xe, xr1, xcs, xss, ssa, sda, ssb, sdb;
+
+    const double eps = 1.0e-15;
+    const double pi = 3.141592653589793;
+    const double c1 = 0.355028053887817;
+    const double c2 = 0.258819403792807;
+    const double sr3 = 1.732050807568877;
+
+    km2 = 0;
+    xa = fabs(x);
+    xq = sqrt(xa);
+    xm = 8.0;
+    if (x > 0.0)
+        xm = 5.0;
+
+    if (x == 0.0) {
+        *ai = c1;
+        *bi = sr3 * c1;
+        *ad = -c2;
+        *bd = sr3 * c2;
+        return;
+    }
+
+    if (xa <= xm) {
+        fx = 1.0;
+        r = 1.0;
+        for (k = 1; k <= 40; k++) {
+            r = r * x / (3.0 * k) * x / (3.0 * k - 1.0) * x;
+            fx += r;
+            if (fabs(r) < fabs(fx) * eps)
+                break;
+        }
+
+        gx = x;
+        r = x;
+        for (k = 1; k <= 40; k++) {
+            r = r * x / (3.0 * k) * x / (3.0 * k + 1.0) * x;
+            gx += r;
+            if (fabs(r) < fabs(gx) * eps)
+                break;
+        }
+
+        *ai = c1 * fx - c2 * gx;
+        *bi = sr3 * (c1 * fx + c2 * gx);
+
+        df = 0.5 * x * x;
+        r = df;
+        for (k = 1; k <= 40; k++) {
+            r = r * x / (3.0 * k) * x / (3.0 * k + 2.0) * x;
+            df += r;
+            if (fabs(r) < fabs(df) * eps)
+                break;
+        }
+
+        dg = 1.0;
+        r = 1.0;
+        for (k = 1; k <= 40; k++) {
+            r = r * x / (3.0 * k) * x / (3.0 * k - 2.0) * x;
+            dg += r;
+            if (fabs(r) < fabs(dg) * eps)
+                break;
+        }
+
+        *ad = c1 * df - c2 * dg;
+        *bd = sr3 * (c1 * df + c2 * dg);
+    } else {
+        km = (int) (24.5 - xa);
+        if (xa < 6.0)
+            km = 14;
+        if (xa > 15.0)
+            km = 10;
+
+        if (x > 0.0) {
+            kmax = km;
+        } else {
+            // Choose cutoffs so that the remainder term in asymptotic
+            // expansion is epsilon size. The X<0 branch needs to be fast
+            // in order to make AIRYZO efficient
+            if (xa > 70.0)
+                km = 3;
+            if (xa > 500.0)
+                km = 2;
+            if (xa > 1000.0)
+                km = 1;
+
+            km2 = km;
+            if (xa > 150.0)
+                km2 = 1;
+            if (xa > 3000.0)
+                km2 = 0;
+            kmax = 2 * km + 1;
+        }
+        xe = xa * xq / 1.5;
+        xr1 = 1.0 / xe;
+        xar = 1.0 / xq;
+        xf = sqrt(xar);
+        rp = 0.5641895835477563;
+        r = 1.0;
+        for (k = 1; k <= kmax; k++) {
+            r = r * (6.0 * k - 1.0) / 216.0 * (6.0 * k - 3.0) / k * (6.0 * k - 5.0) / (2.0 * k - 1.0);
+            ck[k - 1] = r;
+            dk[k - 1] = -(6.0 * k + 1.0) / (6.0 * k - 1.0) * r;
+        }
+
+        if (x > 0.0) {
+            sai = 1.0;
+            sad = 1.0;
+            r = 1.0;
+            for (k = 1; k <= km; k++) {
+                r *= -xr1;
+                sai += ck[k - 1] * r;
+                sad += dk[k - 1] * r;
+            }
+            sbi = 1.0;
+            sbd = 1.0;
+            r = 1.0;
+            for (k = 1; k <= km; k++) {
+                r *= xr1;
+                sbi += ck[k - 1] * r;
+                sbd += dk[k - 1] * r;
+            }
+            xp1 = exp(-xe);
+            *ai = 0.5 * rp * xf * xp1 * sai;
+            *bi = rp * xf / xp1 * sbi;
+            *ad = -0.5 * rp / xf * xp1 * sad;
+            *bd = rp / xf / xp1 * sbd;
+        } else {
+            xcs = cos(xe + pi / 4.0);
+            xss = sin(xe + pi / 4.0);
+            ssa = 1.0;
+            sda = 1.0;
+            r = 1.0;
+            xr2 = 1.0 / (xe * xe);
+            for (k = 1; k <= km; k++) {
+                r *= -xr2;
+                ssa += ck[2 * k - 1] * r;
+                sda += dk[2 * k - 1] * r;
+            }
+            ssb = ck[0] * xr1;
+            sdb = dk[0] * xr1;
+            r = xr1;
+            for (k = 1; k <= km2; k++) {
+                r *= -xr2;
+                ssb += ck[2 * k] * r;
+                sdb += dk[2 * k] * r;
+            }
+
+            *ai = rp * xf * (xss * ssa - xcs * ssb);
+            *bi = rp * xf * (xcs * ssa + xss * ssb);
+            *ad = -rp / xf * (xcs * sda + xss * sdb);
+            *bd = rp / xf * (xss * sda - xcs * sdb);
+        }
+    }
+    return;
+}
+
+inline void airyzo(int nt, int kf, double *xa, double *xb, double *xc, double *xd) {
+
+    // ========================================================
+    // Purpose: Compute the first NT zeros of Airy functions
+    //          Ai(x) and Ai'(x), a and a', and the associated
+    //          values of Ai(a') and Ai'(a); and the first NT
+    //          zeros of Airy functions Bi(x) and Bi'(x), b and
+    //          b', and the associated values of Bi(b') and
+    //          Bi'(b)
+    // Input :  NT    --- Total number of zeros
+    //          KF    --- Function code
+    //                    KF=1 for Ai(x) and Ai'(x)
+    //                    KF=2 for Bi(x) and Bi'(x)
+    // Output:  XA(m) --- a, the m-th zero of Ai(x) or
+    //                    b, the m-th zero of Bi(x)
+    //          XB(m) --- a', the m-th zero of Ai'(x) or
+    //                    b', the m-th zero of Bi'(x)
+    //          XC(m) --- Ai(a') or Bi(b')
+    //          XD(m) --- Ai'(a) or Bi'(b)
+    //                    ( m --- Serial number of zeros )
+    // Routine called: AIRYB for computing Airy functions and
+    //                 their derivatives
+    // =======================================================
+
+    const double pi = 3.141592653589793;
+    int i;
+    double rt = 0.0, rt0, u = 0.0, u1 = 0.0, x, ai, bi, ad, bd, err;
+
+    for (i = 1; i <= nt; ++i) {
+        rt0 = 0.0;
+        if (kf == 1) {
+            u = 3.0 * pi * (4.0 * i - 1) / 8.0;
+            u1 = 1 / (u * u);
+        } else if (kf == 2) {
+            if (i == 1) {
+                rt0 = -1.17371;
+            } else {
+                u = 3.0 * pi * (4.0 * i - 3.0) / 8.0;
+                u1 = 1 / (u * u);
+            }
+        }
+
+        if (rt0 == 0) {
+            // DLMF 9.9.18
+            rt0 = -pow(u * u, 1.0 / 3.0) *
+                  (1.0 +
+                   u1 * (5.0 / 48.0 + u1 * (-5.0 / 36.0 + u1 * (77125.0 / 82944.0 + u1 * (-108056875.0 / 6967296.0)))));
+        }
+
+        while (1) {
+            x = rt0;
+            airyb(x, &ai, &bi, &ad, &bd);
+
+            if (kf == 1) {
+                rt = rt0 - ai / ad;
+            } else if (kf == 2) {
+                rt = rt0 - bi / bd;
+            }
+
+            err = fabs((rt - rt0) / rt);
+            if (err <= 1.0e-12) {
+                break;
+            } else {
+                rt0 = rt;
+            }
+        }
+
+        xa[i - 1] = rt;
+        if (err > 1.0e-14) {
+            airyb(rt, &ai, &bi, &ad, &bd);
+        }
+
+        if (kf == 1) {
+            xd[i - 1] = ad;
+        } else if (kf == 2) {
+            xd[i - 1] = bd;
+        }
+    }
+
+    for (i = 1; i <= nt; ++i) {
+        rt0 = 0.0;
+
+        if (kf == 1) {
+            if (i == 1) {
+                rt0 = -1.01879;
+            } else {
+                u = 3.0 * pi * (4.0 * i - 3.0) / 8.0;
+                u1 = 1 / (u * u);
+            }
+        } else if (kf == 2) {
+            if (i == 1) {
+                rt0 = -2.29444;
+            } else {
+                u = 3.0 * pi * (4.0 * i - 1.0) / 8.0;
+                u1 = 1 / (u * u);
+            }
+        }
+
+        if (rt0 == 0) {
+            // DLMF 9.9.19
+            rt0 = -pow(u * u, 1.0 / 3.0) *
+                  (1.0 + u1 * (-7.0 / 48.0 +
+                               u1 * (35.0 / 288.0 + u1 * (-181223.0 / 207360.0 + u1 * (18683371.0 / 1244160.0)))));
+        }
+
+        while (1) {
+            x = rt0;
+            airyb(x, &ai, &bi, &ad, &bd);
+
+            if (kf == 1) {
+                rt = rt0 - ad / (ai * x);
+            } else if (kf == 2) {
+                rt = rt0 - bd / (bi * x);
+            }
+
+            err = fabs((rt - rt0) / rt);
+            if (err <= 1.0e-12) {
+                break;
+            } else {
+                rt0 = rt;
+            }
+        }
+        xb[i - 1] = rt;
+
+        if (err > 1.0e-14) {
+            airyb(rt, &ai, &bi, &ad, &bd);
+        }
+
+        if (kf == 1) {
+            xc[i - 1] = ai;
+        } else if (kf == 2) {
+            xc[i - 1] = bi;
+        }
+    }
+    return;
+}
+
 template <typename T>
 void itairy(T x, T *apt, T *bpt, T *ant, T *bnt) {
     bool x_signbit = std::signbit(x);
