+++ scipy/special/special/specfun/specfun.h
@@ -80,10 +80,6 @@ void bjndd(double, int, double *, double *, double *);
 
 void cerzo(int, std::complex<double> *);
 
-std::complex<double> cerror(std::complex<double>);
-std::complex<double> cchg(double, double, std::complex<double>);
-std::complex<double> cgama(std::complex<double>, int);
-
 void clpmn(std::complex<double>, int, int, int, std::complex<double> *, std::complex<double> *);
 void clpn(int, std::complex<double>, std::complex<double> *, std::complex<double> *);
 
@@ -102,9 +98,6 @@ double cv0(double, double, double);
 double cvf(int, int, double, double, int);
 double cvql(int, int, double);
 double cvqm(int, double);
-double dvla(double, double);
-double dvsa(double, double);
-void fcszo(int, int, std::complex<double> *);
 double gaih(double);
 double gam0(double);
 double gamma2(double);
@@ -131,314 +124,6 @@ void sphj(T, int, int *, T *, T *);
 
 template <typename T>
 void sphy(T, int, int *, T *, T *);
-
-double vvla(double, double);
-double vvsa(double, double);
-
-
-inline void airyb(double x, double* ai, double* bi, double* ad, double* bd) {
-
-    // =======================================================
-    // Purpose: Compute Airy functions and their derivatives
-    // Input:   x  --- Argument of Airy function
-    // Output:  AI --- Ai(x)
-    //          BI --- Bi(x)
-    //          AD --- Ai'(x)
-    //          BD --- Bi'(x)
-    // =======================================================
-
-    int k, km, km2, kmax;
-    double ck[52], dk[52];
-    double xa, xq, xm, fx, r, gx, df, dg, sai, sad, sbi, sbd, xp1, xr2;
-    double xf, rp, xar, xe, xr1, xcs, xss, ssa, sda, ssb, sdb;
-
-    const double eps = 1.0e-15;
-    const double pi = 3.141592653589793;
-    const double c1 = 0.355028053887817;
-    const double c2 = 0.258819403792807;
-    const double sr3 = 1.732050807568877;
-
-    km2 = 0;
-    xa = fabs(x);
-    xq = sqrt(xa);
-    xm = 8.0;
-    if (x > 0.0) xm = 5.0;
-
-    if (x == 0.0) {
-        *ai = c1;
-        *bi = sr3 * c1;
-        *ad = -c2;
-        *bd = sr3 * c2;
-        return;
-    }
-
-    if (xa <= xm) {
-        fx = 1.0;
-        r = 1.0;
-        for (k = 1; k <= 40; k++) {
-            r = r * x / (3.0 * k) * x / (3.0 * k - 1.0) * x;
-            fx += r;
-            if (fabs(r) < fabs(fx) * eps) break;
-        }
-
-        gx = x;
-        r = x;
-        for (k = 1; k <= 40; k++) {
-            r = r * x / (3.0 * k) * x / (3.0 * k + 1.0) * x;
-            gx += r;
-            if (fabs(r) < fabs(gx) * eps) break;
-        }
-
-        *ai = c1 * fx - c2 * gx;
-        *bi = sr3 * (c1 * fx + c2 * gx);
-
-        df = 0.5 * x * x;
-        r = df;
-        for (k = 1; k <= 40; k++) {
-            r = r * x / (3.0 * k) * x / (3.0 * k + 2.0) * x;
-            df += r;
-            if (fabs(r) < fabs(df) * eps) break;
-        }
-
-        dg = 1.0;
-        r = 1.0;
-        for (k = 1; k <= 40; k++) {
-            r = r * x / (3.0 * k) * x / (3.0 * k - 2.0) * x;
-            dg += r;
-            if (fabs(r) < fabs(dg) * eps) break;
-        }
-
-        *ad = c1 * df - c2 * dg;
-        *bd = sr3 * (c1 * df + c2 * dg);
-    } else {
-        km = (int)(24.5 - xa);
-        if (xa < 6.0) km = 14;
-        if (xa > 15.0) km = 10;
-
-        if (x > 0.0) {
-            kmax = km;
-        } else {
-            // Choose cutoffs so that the remainder term in asymptotic
-            // expansion is epsilon size. The X<0 branch needs to be fast
-            // in order to make AIRYZO efficient
-            if (xa > 70.0) km = 3;
-            if (xa > 500.0) km = 2;
-            if (xa > 1000.0) km = 1;
-
-            km2 = km;
-            if (xa > 150.0) km2 = 1;
-            if (xa > 3000.0) km2 = 0;
-            kmax = 2 * km + 1;
-        }
-        xe = xa * xq / 1.5;
-        xr1 = 1.0 / xe;
-        xar = 1.0 / xq;
-        xf = sqrt(xar);
-        rp = 0.5641895835477563;
-        r = 1.0;
-        for (k = 1; k <= kmax; k++) {
-            r = r * (6.0 * k - 1.0) / 216.0 * (6.0 * k - 3.0)
-                / k * (6.0 * k - 5.0) / (2.0 * k - 1.0);
-            ck[k - 1] = r;
-            dk[k - 1] = -(6.0 * k + 1.0) / (6.0 * k - 1.0) * r;
-        }
-
-        if (x > 0.0) {
-            sai = 1.0;
-            sad = 1.0;
-            r = 1.0;
-            for (k = 1; k <= km; k++) {
-                r *= -xr1;
-                sai += ck[k - 1] * r;
-                sad += dk[k - 1] * r;
-            }
-            sbi = 1.0;
-            sbd = 1.0;
-            r = 1.0;
-            for (k = 1; k <= km; k++) {
-                r *= xr1;
-                sbi += ck[k - 1] * r;
-                sbd += dk[k - 1] * r;
-            }
-            xp1 = exp(-xe);
-            *ai = 0.5 * rp * xf * xp1 * sai;
-            *bi = rp * xf / xp1 * sbi;
-            *ad = -0.5 * rp / xf * xp1 * sad;
-            *bd = rp / xf / xp1 * sbd;
-        } else {
-            xcs = cos(xe + pi / 4.0);
-            xss = sin(xe + pi / 4.0);
-            ssa = 1.0;
-            sda = 1.0;
-            r = 1.0;
-            xr2 = 1.0 / (xe * xe);
-            for (k = 1; k <= km; k++) {
-                r *= -xr2;
-                ssa += ck[2 * k - 1] * r;
-                sda += dk[2 * k - 1] * r;
-            }
-            ssb = ck[0] * xr1;
-            sdb = dk[0] * xr1;
-            r = xr1;
-            for (k = 1; k <= km2; k++) {
-                r *= -xr2;
-                ssb += ck[2 * k] * r;
-                sdb += dk[2 * k] * r;
-            }
-
-            *ai = rp * xf * (xss * ssa - xcs * ssb);
-            *bi = rp * xf * (xcs * ssa + xss * ssb);
-            *ad = -rp / xf * (xcs * sda + xss * sdb);
-            *bd = rp / xf * (xss * sda - xcs * sdb);
-        }
-    }
-    return;
-}
-
-
-inline void airyzo(int nt, int kf, double *xa, double *xb, double *xc, double *xd) {
-
-    // ========================================================
-    // Purpose: Compute the first NT zeros of Airy functions
-    //          Ai(x) and Ai'(x), a and a', and the associated
-    //          values of Ai(a') and Ai'(a); and the first NT
-    //          zeros of Airy functions Bi(x) and Bi'(x), b and
-    //          b', and the associated values of Bi(b') and
-    //          Bi'(b)
-    // Input :  NT    --- Total number of zeros
-    //          KF    --- Function code
-    //                    KF=1 for Ai(x) and Ai'(x)
-    //                    KF=2 for Bi(x) and Bi'(x)
-    // Output:  XA(m) --- a, the m-th zero of Ai(x) or
-    //                    b, the m-th zero of Bi(x)
-    //          XB(m) --- a', the m-th zero of Ai'(x) or
-    //                    b', the m-th zero of Bi'(x)
-    //          XC(m) --- Ai(a') or Bi(b')
-    //          XD(m) --- Ai'(a) or Bi'(b)
-    //                    ( m --- Serial number of zeros )
-    // Routine called: AIRYB for computing Airy functions and
-    //                 their derivatives
-    // =======================================================
-
-    const double pi = 3.141592653589793;
-    int i;
-    double rt = 0.0, rt0, u = 0.0, u1 = 0.0, x, ai, bi, ad, bd, err;
-
-    for (i = 1; i <= nt; ++i) {
-        rt0 = 0.0;
-        if (kf == 1) {
-            u = 3.0 * pi * (4.0 * i - 1) / 8.0;
-            u1 = 1 / (u * u);
-        } else if (kf == 2) {
-            if (i == 1) {
-                rt0 = -1.17371;
-            } else {
-                u = 3.0 * pi * (4.0 * i - 3.0) / 8.0;
-                u1 = 1 / (u * u);
-            }
-        }
-
-        if (rt0 == 0) {
-            // DLMF 9.9.18
-            rt0 = -pow(u * u, 1.0 / 3.0) * (
-                    1.0
-                    + u1 * (5.0 / 48.0
-                    + u1 * (-5.0 / 36.0
-                    + u1 * (77125.0 / 82944.0
-                    + u1 * (-108056875.0 / 6967296.0)))));
-        }
-
-        while (1) {
-            x = rt0;
-            airyb(x, &ai, &bi, &ad, &bd);
-
-            if (kf == 1) {
-                rt = rt0 - ai / ad;
-            } else if (kf == 2) {
-                rt = rt0 - bi / bd;
-            }
-
-            err = fabs((rt - rt0) / rt);
-            if (err <= 1.0e-12) {
-                break;
-            } else {
-                rt0 = rt;
-            }
-        }
-
-        xa[i - 1] = rt;
-        if (err > 1.0e-14) {
-            airyb(rt, &ai, &bi, &ad, &bd);
-        }
-
-        if (kf == 1) {
-            xd[i - 1] = ad;
-        } else if (kf == 2) {
-            xd[i - 1] = bd;
-        }
-    }
-
-    for (i = 1; i <= nt; ++i) {
-        rt0 = 0.0;
-
-        if (kf == 1) {
-            if (i == 1) {
-                rt0 = -1.01879;
-            } else {
-                u = 3.0 * pi * (4.0 * i - 3.0) / 8.0;
-                u1 = 1 / (u * u);
-            }
-        } else if (kf == 2) {
-            if (i == 1) {
-                rt0 = -2.29444;
-            } else {
-                u = 3.0 * pi * (4.0 * i - 1.0) / 8.0;
-                u1 = 1 / (u * u);
-            }
-        }
-
-        if (rt0 == 0) {
-            // DLMF 9.9.19
-            rt0 = -pow(u * u, 1.0 / 3.0) * (
-                    1.0
-                    + u1 * (-7.0 / 48.0
-                    + u1 * (35.0 / 288.0
-                    + u1 * (-181223.0 / 207360.0
-                    + u1 * (18683371.0 / 1244160.0)))));
-        }
-
-        while (1) {
-            x = rt0;
-            airyb(x, &ai, &bi, &ad, &bd);
-
-            if (kf == 1) {
-                rt = rt0 - ad / (ai * x);
-            } else if (kf == 2) {
-                rt = rt0 - bd / (bi * x);
-            }
-
-            err = fabs((rt - rt0) / rt);
-            if (err <= 1.0e-12) {
-                break;
-            } else {
-                rt0 = rt;
-            }
-        }
-        xb[i - 1] = rt;
-
-        if (err > 1.0e-14) {
-            airyb(rt, &ai, &bi, &ad, &bd);
-        }
-
-        if (kf == 1) {
-            xc[i - 1] = ai;
-        } else if (kf == 2) {
-            xc[i - 1] = bi;
-        }
-    }
-    return;
-}
-
 
 template <typename T>
 void aswfa(T x, int m, int n, T c, int kd, T cv, T *s1f, T *s1d) {
@@ -858,179 +543,6 @@ inline void cerzo(int nt, std::complex<double> *zo) {
     }
     return;
 }
-
-
-inline void cfc(std::complex<double> z, std::complex<double> *zf, std::complex<double> *zd) {
-
-    // =========================================================
-    // Purpose: Compute complex Fresnel integral C(z) and C'(z)
-    // Input :  z --- Argument of C(z)
-    // Output:  ZF --- C(z)
-    //          ZD --- C'(z)
-    // =========================================================
-
-    int k, m;
-    double wa0, wa;
-    std::complex<double> c, cr, cf, cf0, cf1, cg, d;
-    const double eps = 1.0e-14;
-    const double pi = 3.141592653589793;
-
-    double w0 = std::abs(z);
-    std::complex<double> zp = 0.5 * pi * z * z;
-    std::complex<double> zp2 = zp * zp;
-    std::complex<double> z0 = 0.0;
-
-    if (z == z0) {
-        c = z0;
-    } else if (w0 <= 2.5) {
-        cr = z;
-        c = cr;
-        wa0 = 0.0;
-        for (k = 1; k <= 80; k++) {
-            cr = -0.5*cr*(4.0*k - 3.0)/static_cast<double>(k)/(2.0*k - 1.0)/(4.0*k + 1.0)*zp2;
-            c += cr;
-            wa = std::abs(c);
-            if ((fabs((wa - wa0) / wa) < eps) && (k > 10)) {
-                *zf = c;
-                *zd = std::cos(0.5*pi*z*z);
-                return;
-            }
-            wa0 = wa;
-        }
-    } else if ((w0 > 2.5) && (w0 < 4.5)) {
-        m = 85;
-        c = z0;
-        cf1 = z0;
-        cf0 = 1.0e-100;
-        for (k = m; k >= 0; k--) {
-            cf = (2.0*k + 3.0)*cf0/zp - cf1;
-            if (k % 2 == 0) { c += cf; }
-            cf1 = cf0;
-            cf0 = cf;
-        }
-        c *= 2.0/(pi*z)*std::sin(zp)/cf;
-    } else {
-        // See comment at CFS(), use C(z) = iC(-iz)
-        if ((z.imag() > -z.real()) && (z.imag() <= z.real())) {
-            // right quadrant
-            d = 0.5;
-        } else if ((z.imag() > z.real()) && (z.imag() >= -z.real())) {
-            // upper quadrant
-            d = std::complex<double>(0, 0.5);
-        } else if ((z.imag() < -z.real()) && (z.imag() >= z.real())) {
-            // left quadrant
-            d = -0.5;
-        } else {
-            d = std::complex<double>(0, -0.5);
-        }
-        cr = 1.0;
-        cf = 1.0;
-        for (k = 1; k <= 20; k++) {
-            cr = -0.25*cr*(4.0*k - 1.0)*(4.0*k - 3.0)/zp2;
-            cf += cr;
-        }
-        cr = 1.0/(pi*z*z);
-        cg = cr;
-        for (k = 1; k <= 12; k++) {
-            cr = -0.25*cr*(4.0*k + 1.0)*(4.0*k - 1.0)/zp2;
-            cg += cr;
-        }
-        c = d + (cf*std::sin(zp) - cg*std::cos(zp))/(pi*z);
-    }
-    *zf = c;
-    *zd = std::cos(0.5*pi*z*z);
-    return;
-}
-
-
-inline void cfs(std::complex<double> z, std::complex<double> *zf, std::complex<double> *zd) {
-
-    // =========================================================
-    // Purpose: Compute complex Fresnel Integral S(z) and S'(z)
-    // Input :  z  --- Argument of S(z)
-    // Output:  ZF --- S(z)
-    //          ZD --- S'(z)
-    // =========================================================
-
-    int k, m;
-    double wb0, wb;
-    std::complex<double> s, cr, cf, cf0, cf1, cg, d;
-    const double eps = 1.0e-14;
-    const double pi = 3.141592653589793;
-
-    double w0 = std::abs(z);
-    std::complex<double> zp = 0.5 * pi * z * z;
-    std::complex<double> zp2 = zp * zp;
-    std::complex<double> z0 = 0.0;
-
-    if (z == z0) {
-        s = z0;
-    } else if (w0 <= 2.5) {
-        s = z * zp / 3.0;
-        cr = s;
-        wb0 = 0.0;
-        for (k = 1; k <= 80; k++) {
-            cr = -0.5 * cr * (4.0 * k - 1.0) / static_cast<double>(k) / (2.0 * k + 1.0) / (4.0 * k + 3.0) * zp2;
-            s += cr;
-            wb = std::abs(s);
-            if ((fabs(wb - wb0) < eps) && (k > 10)) {
-                    *zf = s;
-                    *zd = std::sin(0.5*pi*z*z);
-                    return;
-            }
-            wb0 = wb;
-        }
-    } else if ((w0 > 2.5) && (w0 < 4.5)) {
-        m = 85;
-        s = z0;
-        cf1 = z0;
-        cf0 = 1.0e-100;
-        for (k = m; k >= 0; k--) {
-            cf = (2.0*k + 3.0)*cf0/zp - cf1;
-            if (k % 2 == 1) { s += cf; }
-            cf1 = cf0;
-            cf0 = cf;
-        }
-        s = 2.0/(pi * z)*std::sin(zp)/cf*s;
-    } else {
-        // Auxiliary functions f(z) and g(z) can be computed using an
-        // asymptotic expansion in the right quadrant |arg(z)| <= pi/4, not pi/2
-        // as sometimes suggested. Use the symmetry S(z) = -iS(-iz).
-        // Interestingly, most of the expansion code is the same across
-        // the quadrants. (The forth power in Z is the equalizer here.)
-        // Only one constant has to be adapted.
-        if ((z.imag() > -z.real()) && (z.imag() <= z.real())) {
-            // right quadrant
-            d = 0.5;
-        } else if ((z.imag() > z.real()) && (z.imag() >= -z.real())) {
-            // upper quadrant
-            d = std::complex<double>(0, -0.5);
-        } else if ((z.imag() < -z.real()) && (z.imag() >= z.real())) {
-            // left quadrant
-            d = -0.5;
-        } else {
-            d = std::complex<double>(0, 0.5);
-        }
-        cr = 1.0;
-        cf = 1.0;
-        for (k = 1; k <= 20; k++) {
-            cr = -0.25*cr*(4.0*k - 1.0)*(4.0*k - 3.0)/zp2;
-            cf += cr;
-        }
-        cr = 1.0;
-        cg = 1.0;
-        for (k = 1; k <= 12; k++) {
-            cr = -0.25*cr*(4.0*k + 1.0)*(4.0*k - 1.0)/zp2;
-            cg += cr;
-        }
-        cg = cg/(pi*z*z);
-        s = d - (cf*std::cos(zp) + cg*std::sin(zp))/(pi*z);
-    }
-    *zf = s;
-    *zd = std::sin(0.5*pi*z*z);
-    return;
-}
-
 
 inline std::complex<double> cchg(double a, double b, std::complex<double> z) {
 
@@ -2827,94 +2339,8 @@ inline void cyzo(int nt, int kf, int kc, std::complex<double> *zo, std::complex<
 }
 
 
-inline double dvla(double x, double va) {
-
-    // ====================================================
-    // Purpose: Compute parabolic cylinder functions Dv(x)
-    //          for large argument
-    // Input:   x  --- Argument
-    //          va --- Order
-    // Output:  PD --- Dv(x)
-    // Routines called:
-    //       (1) VVLA for computing Vv(x) for large |x|
-    //       (2) GAMMA2 for computing Г(x)
-    // ====================================================
-
-    int k;
-    double ep, a0, gl, pd, r, vl, x1;
-
-    const double pi=3.141592653589793;
-    const double eps=1.0e-12;
-    ep = exp(-.25*x*x);
-    a0 = pow(fabs(x), va)*ep;
-    r=1.0;
-    pd=1.0;
-    for (k = 1; k <= 16; k++) {
-        r = -0.5*r*(2.0*k-va-1.0)*(2.0*k-va-2.0)/(k*x*x);
-        pd += r;
-        if (fabs(r/pd) < eps) { break; }
-    }
-    pd *= a0;
-    if (x < 0.0) {
-        x1 = -x;
-        vl = vvla(x1, va);
-        gl = gamma2(-va);
-        pd = pi*vl/gl + cos(pi*va)*pd;
-    }
-    return pd;
-}
 
 
-inline double dvsa(double x, double va) {
-
-    // ===================================================
-    // Purpose: Compute parabolic cylinder function Dv(x)
-    //          for small argument
-    // Input:   x  --- Argument
-    //          va --- Order
-    // Output:  PD --- Dv(x)
-    // Routine called: GAMMA2 for computing Г(x)
-    // ===================================================
-
-    int m;
-    double ep, a0, va0, pd, ga0, g1, g0, r, vm, gm, r1, vt;
-
-    const double pi = 3.141592653589793;
-    const double eps = 1.0e-15;
-    const double sq2 = sqrt(2);
-
-    ep = exp(-0.25*x*x);
-    va0 = 0.5*(1.0 - va);
-    if (va == 0.0) {
-        pd = ep;
-    } else {
-        if (x == 0.0) {
-            if ((va0 <= 0.0) && (va0 == (int)va0)) {
-                pd = 0.0;
-            } else {
-                ga0 = gamma2(va0);
-                pd = sqrt(pi)/(pow(2.0, -0.5*va)*ga0);
-            }
-        } else {
-            g1 = gamma2(-va);
-            a0 = pow(2.0, -0.5*va - 1.0)*ep/g1;
-            vt = -0.5*va;
-            g0 = gamma2(vt);
-            pd = g0;
-            r = 1.0;
-            for (m = 1; m <= 250; m++) {
-                vm = 0.5*(m-va);
-                gm = gamma2(vm);
-                r = -r*sq2*x/m;
-                r1 = gm*r;
-                pd += r1;
-                if (fabs(r1) < fabs(pd)*eps) { break; }
-            }
-            pd *= a0;
-        }
-    }
-    return pd;
-}
 
 
 template <typename T>
@@ -3388,75 +2814,6 @@ L70:
 }
 
 
-inline void fcszo(int kf, int nt, std::complex<double> *zo) {
-
-    // ===============================================================
-    // Purpose: Compute the complex zeros of Fresnel integral C(z)
-    //          or S(z) using modified Newton's iteration method
-    // Input :  KF  --- Function code
-    //                  KF=1 for C(z) or KF=2 for S(z)
-    //          NT  --- Total number of zeros
-    // Output:  ZO(L) --- L-th zero of C(z) or S(z)
-    // Routines called:
-    //      (1) CFC for computing Fresnel integral C(z)
-    //      (2) CFS for computing Fresnel integral S(z)
-    // ==============================================================
-
-
-    int i, j, it, nr;
-    double psq, px, py, w, w0;
-    std::complex<double> z, zp, zf, zd, zfd, zgd, zq, zw;
-    const double pi = 3.141592653589793;
-    psq = 0.0;
-    w = 0.0;
-
-    for (nr = 1; nr <= nt; ++nr) {
-        if (kf == 1)
-            psq = sqrt(4.0 * nr - 1.0);
-        if (kf == 2)
-            psq = 2.0 * sqrt(nr);
-
-        px = psq - log(pi * psq) / (pi * pi * psq * psq * psq);
-        py = log(pi * psq) / (pi * psq);
-        z = std::complex<double>(px, py);
-
-        if (kf == 2) {
-            if (nr == 2) { z = std::complex<double>(2.8334, 0.2443); }
-            if (nr == 3) { z = std::complex<double>(3.4674, 0.2185); }
-            if (nr == 4) { z = std::complex<double>(4.0025, 0.2008); }
-        }
-
-        it = 0;
-        do {
-            it++;
-            if (kf == 1) { cfc(z, &zf, &zd); }
-            if (kf == 2) { cfs(z, &zf, &zd); }
-
-            zp = 1.0;
-            for (i = 1; i < nr; i++)
-                zp *= (z - zo[i - 1]);
-
-            zfd = zf / zp;
-            zq = 0.0;
-            for (i = 1; i < nr ; i++) {
-                zw = 1.0;
-                for (j = 1; j < nr; j++) {
-                    if (j == i) { continue; }
-                    zw *= (z - zo[j - 1]);
-                }
-                zq += zw;
-            }
-            zgd = (zd - zq * zfd) / zp;
-            z -= zfd / zgd;
-            w0 = w;
-            w = std::abs(z);
-        } while ((it <= 50) && (fabs((w - w0) / w) > 1.0e-12));
-        zo[nr - 1] = z;
-    }
-    return;
-}
-
-
 inline double gaih(double x) {
 
     // =====================================================
@@ -3917,311 +3274,6 @@ inline std::complex<double> hygfz(double a, double b, double c, std::complex<dou
     if (k > 150) { *isfer = 5; }
     return zhf;
 }
-
-
-template <typename T>
-void itika(T x, T *ti, T *tk) {
-
-    // =======================================================
-    // Purpose: Integrate modified Bessel functions I0(t) and
-    //          K0(t) with respect to t from 0 to x
-    // Input :  x  --- Upper limit of the integral  ( x ≥ 0 )
-    // Output:  TI --- Integration of I0(t) from 0 to x
-    //          TK --- Integration of K0(t) from 0 to x
-    // =======================================================
-
-    int k;
-    T rc1, rc2, e0, b1, b2, rs, r, tw, x2;
-    static const T a[10] = {
-        0.625, 1.0078125, 2.5927734375, 9.1868591308594,
-        4.1567974090576e+1, 2.2919635891914e+2,1.491504060477e+3,
-        1.1192354495579e+4, 9.515939374212e+4,9.0412425769041e+5
-    };
-    const T pi = 3.141592653589793;
-    const T el = 0.5772156649015329;
-
-    if (x == 0.0) {
-        *ti = 0.0;
-        *tk = 0.0;
-        return;
-    } else if (x < 20.0) {
-        x2 = x*x;
-        *ti = 1.0;
-        r = 1.0;
-        for (k = 1; k <= 50; k++) {
-            r = 0.25 * r *(2 * k - 1.0) / (2 * k + 1.0) / (k * k) * x2;
-            *ti += r;
-            if (fabs(r/(*ti)) < 1.0e-12) { break; }
-        }
-        *ti *= x;
-    } else {
-        x2 = 0.0;
-        *ti = 1.0;
-        r = 1.0;
-        for (k = 1; k <= 10; k++) {
-            r = r/x;
-            *ti += a[k-1]*r;
-        }
-        rc1 = 1.0 / sqrt(2.0*pi*x);
-        *ti = rc1 * exp(x) * (*ti);
-    }
-
-    if (x < 12.0) {
-        e0 = el + log(x / 2.0);
-        b1 = 1.0 - e0;
-        b2 = 0.0;
-        rs = 0.0;
-        r = 1.0;
-        tw = 0.0;
-        for (k = 1; k <= 50; k++) {
-            r = 0.25*r*(2*k - 1.0) / (2*k + 1.0) / (k*k)*x2;
-            b1 += r * (1.0 / (2 * k + 1) - e0);
-            rs += 1.0 / k;
-            b2 += r * rs;
-            *tk = b1 + b2;
-            if (fabs((*tk - tw) / (*tk)) < 1.0e-12) { break; }
-            tw = *tk;
-        }
-        *tk *= x;
-    } else {
-        *tk = 1.0;
-        r = 1.0;
-        for (k = 1; k <= 10; k++) {
-            r = -r / x;
-            *tk = *tk + a[k-1]*r;
-        }
-        rc2 = sqrt(pi/(2.0*x));
-        *tk = pi/2.0 - rc2*(*tk)*exp(-x);
-    }
-    return;
-}
-
-
-template <typename T>
-void itjya(T x, T *tj, T *ty) {
-    int k;
-    T a[18], a0, a1, af, bf, bg, r, r2, rc, rs, ty1, ty2, x2, xp;
-    const T pi = 3.141592653589793;
-    const T el = 0.5772156649015329;
-    const T eps = 1.0e-12;
-
-    if (x == 0.0) {
-        *tj = 0.0;
-        *ty = 0.0;
-    } else if (x <= 20.0) {
-        x2 = x*x;
-        *tj = x;
-        r = x;
-        for (k = 1; k < 61; k++) {
-            r = -0.25*r*(2.0*k - 1.0) / (2.0*k + 1.0) / (k*k)*x2;
-            *tj += r;
-            if (fabs(r) < fabs(*tj)*eps) { break; }
-        }
-        ty1 = (el + log(x/2.0)) * (*tj);
-        rs = 0.0;
-        ty2 = 1.0;
-        r = 1.0;
-        for (k = 1; k < 61; k++) {
-            r = -0.25*r*(2.0*k - 1.0) / (2.0*k + 1.0) / (k*k) * x2;
-            rs += 1.0 / k;
-            r2 = r * (rs + 1.0 / (2.0*k + 1.0));
-            ty2 += r2;
-            if (fabs(r2) < fabs(ty2)*eps) { break; }
-        }
-        *ty = (ty1 - x*ty2)*2.0/pi;
-    } else {
-        a0 = 1.0;
-        a1 = 5.0 / 8.0;
-        a[0] = a1;
-
-        for (int k = 1; k <= 16; k++) {
-            af = ((1.5*(k+0.5)*(k+5.0/6.0)*a1-0.5*(k+0.5)*(k+0.5)*(k-0.5)*a0))/(k+1.0);
-            a[k] = af;
-            a0 = a1;
-            a1 = af;
-        }
-        bf = 1.0;
-        r = 1.0;
-        for (int k = 1; k <= 8; k++) {
-            r = -r / (x*x);
-            bf += a[2*k-1] * r;
-        }
-        bg = a[0] / x;
-        r = 1.0 / x;
-        for (int k = 1; k <= 8; k++) {
-            r = -1.0/(x*x);
-            bg += a[2*k] * r;
-        }
-        xp = x + 0.25 * pi;
-        rc = sqrt(2.0 / (pi * x));
-        *tj = 1.0 - rc*(bf*cos(xp) + bg*sin(xp));
-        *ty = rc*(bg*cos(xp) - bf*sin(xp));
-    }
-    return;
-}
-
-
-template <typename T>
-void ittika(T x, T *tti, T *ttk) {
-
-    // =========================================================
-    // Purpose: Integrate [I0(t)-1]/t with respect to t from 0
-    //          to x, and K0(t)/t with respect to t from x to ∞
-    // Input :  x   --- Variable in the limits  ( x ≥ 0 )
-    // Output:  TTI --- Integration of [I0(t)-1]/t from 0 to x
-    //          TTK --- Integration of K0(t)/t from x to ∞
-    // =========================================================
-
-    int k;
-    T b1, e0, r, r2, rc, rs;
-    const T pi = 3.141592653589793;
-    const T el = 0.5772156649015329;
-    static const T c[8] = {
-        1.625, 4.1328125, 1.45380859375, 6.553353881835, 3.6066157150269,
-        2.3448727161884, 1.7588273098916,1.4950639538279
-    };
-
-    if (x == 0.0) {
-        *tti = 0.0;
-        *ttk = 1.0e+300;
-        return;
-    }
-    if (x < 40.0) {
-        *tti = 1.0;
-        r = 1.0;
-        for (int k = 2; k <= 50; k++) {
-            r = 0.25*r*(k - 1.0) / (k*k*k)*x*x;
-            *tti += r;
-            if (fabs(r / (*tti)) < 1.0e-12) { break; }
-        }
-        *tti *= 0.125*x*x;
-    } else {
-        *tti = 1.0;
-        r = 1.0;
-        for (int k = 1; k <= 8; k++) {
-            r = r / x;
-            *tti += c[k-1] * r;
-        }
-        rc = x * sqrt(2.0 * pi * x);
-        *tti = (*tti) * exp(x) / rc;
-    }
-    if (x <= 12.0) {
-        e0 = (0.5 * log(x / 2.0) + el) * log(x / 2.0)
-                + pi * pi / 24.0 + 0.5 * el * el;
-        b1 = 1.5 - (el + log(x / 2.0));
-        rs = 1.0;
-        r = 1.0;
-        for (k = 2; k <= 50; k++) {
-            r = 0.25*r*(k - 1.0) / (k*k*k)*x*x;
-            rs += 1.0 / k;
-            r2 = r * (rs + 1.0 / (2.0*k) - (el + log(x / 2.0)));
-            b1 += r2;
-            if (fabs(r2 / b1) < 1.0e-12) { break; }
-        }
-        *ttk = e0 - 0.125*x*x*b1;
-    } else {
-        *ttk = 1.0;
-        r = 1.0;
-        for (k = 1; k <= 8; k++) {
-            r = -r / x;
-            *ttk += c[k-1] * r;
-        }
-        rc = x*sqrt(2.0 / (pi * x));
-        *ttk = (*ttk) * exp(-x) / rc;
-    }
-    return;
-}
-
-
-template <typename T>
-void ittjya(T x, T *ttj, T *tty) {
-
-    // =========================================================
-    // Purpose: Integrate [1-J0(t)]/t with respect to t from 0
-    //          to x, and Y0(t)/t with respect to t from x to ∞
-    // Input :  x   --- Variable in the limits  ( x ≥ 0 )
-    // Output:  TTJ --- Integration of [1-J0(t)]/t from 0 to x
-    //          TTY --- Integration of Y0(t)/t from x to ∞
-    // =========================================================
-
-    int k, l;
-    T a0, bj0, bj1, by0, by1, e0, b1, g0, g1, px, qx, r, r0, r1, r2, rs, t, vt, xk;
-    const T pi = 3.141592653589793;
-    const T el = 0.5772156649015329;
-
-    if (x == 0.0) {
-        *ttj = 0.0;
-        *tty = -1.0e+300;
-    } else if (x <= 20.0) {
-        *ttj = 1.0;
-        r = 1.0;
-        for (k = 2; k <= 100; k++) {
-            r = -0.25 * r * (k - 1.0) / (k*k*k)*x*x;
-            *ttj += r;
-            if (fabs(r) < fabs(*ttj) * 1.0e-12) { break; }
-        }
-        *ttj *= 0.125*x*x;
-        e0 = 0.5*(pi*pi/6.0 - el*el) - (0.5*log(x / 2.0) + el)*log(x/2.0);
-        b1 = el + log(x / 2.0) - 1.5;
-        rs = 1.0;
-        r = -1.0;
-        for (k = 2; k <= 100; k++) {
-            r = -0.25 * r * (k - 1.0) / (k*k*k)*x*x;
-            rs = rs + 1.0 / k;
-            r2 = r * (rs + 1.0/(2.0 * k) - (el + log(x/2.0)));
-            b1 = b1 + r2;
-            if (fabs(r2) < fabs(b1) * 1.0e-12) { break; }
-        }
-        *tty = 2.0/pi*(e0 + 0.125 * x * x * b1);
-    } else {
-        a0 = sqrt(2.0 / (pi * x));
-        bj0 = 0.0;
-        by0 = 0.0;
-        bj1 = 0.0;
-        for (l = 0; l <= 1; l++) {
-            vt = 4.0 * l * l;
-            px = 1.0;
-            r = 1.0;
-            for (k = 1; k <= 14; k++) {
-                r = -0.0078125*r*(vt-pow((4.0*k-3.0), 2))/(x*k)*(vt-pow((4.0*k-1.0), 2))/((2.0*k-1.0)*x);
-                px += r;
-                if (fabs(r) < fabs(px) * 1.0e-12) { break; }
-            }
-            qx = 1.0;
-            r = 1.0;
-            for (k = 1; k <= 14; k++) {
-                r = -0.0078125*r*(vt-pow((4.0*k-1.0), 2))/(x*k)*(vt-pow((4.0*k+1.0), 2))/((2.0*k+1.0)*x);
-                qx += r;
-                if (fabs(r) < fabs(qx) * 1.0e-12) { break; }
-            }
-            qx = 0.125 * (vt - 1.0) / x * qx;
-            xk = x - (0.25 + 0.5 * l)*pi;
-            bj1 = a0*(px*cos(xk) - qx*sin(xk));
-            by1 = a0*(px*sin(xk) + qx*cos(xk));
-            if (l == 0) {
-                bj0 = bj1;
-                by0 = by1;
-            }
-        }
-        t = 2.0 / x;
-        g0 = 1.0;
-        r0 = 1.0;
-        for (k = 1; k <= 10; k++) {
-            r0 = -r0*k*k*t*t;
-            g0 += r0;
-        }
-        g1 = 1.0;
-        r1 = 1.0;
-        for (k = 1; k <= 10; k++) {
-            r1 = -r1 * k * (k + 1.0) * t * t;
-            g1 += r1;
-        }
-        *ttj = 2.0*g1*bj0 / (x*x) - g0*bj1 / x + el + log(x / 2.0);
-        *tty = 2.0*g1*by0 / (x*x) - g0*by1 / x;
-    }
-    return;
-}
-
 
 inline void jdzo(int nt, double *zo, int *n, int *m, int *p) {
 
@@ -4728,240 +3780,6 @@ L25:
         goto L25;
     }
     return;
-}
-
-
-template <typename T>
- void klvna(T x, T *ber, T *bei, T *ger, T *gei,
-                   T *der, T *dei, T *her, T *hei) {
-
-    // ======================================================
-    // Purpose: Compute Kelvin functions ber x, bei x, ker x
-    //          and kei x, and their derivatives  ( x > 0 )
-    // Input :  x   --- Argument of Kelvin functions
-    // Output:  BER --- ber x
-    //          BEI --- bei x
-    //          GER --- ker x
-    //          GEI --- kei x
-    //          DER --- ber'x
-    //          DEI --- bei'x
-    //          HER --- ker'x
-    //          HEI --- kei'x
-    // ================================================
-
-    int k, km, m;
-    T gs, r, x2, x4, pp1, pn1, qp1, qn1, r1, pp0, pn0, qp0, qn0, r0,\
-           fac, xt, cs, ss, xd, xe1, xe2, xc1, xc2, cp0, cn0, sp0, sn0, rc, rs;
-    const T pi = 3.141592653589793;
-    const T el = 0.5772156649015329;
-    const T eps = 1.0e-15;
-
-    if (x == 0.0) {
-        *ber = 1.0;
-        *bei = 0.0;
-        *ger = 1.0e+300;
-        *gei = -0.25 * pi;
-        *der = 0.0;
-        *dei = 0.0;
-        *her = -1.0e+300;
-        *hei = 0.0;
-        return;
-    }
-
-    x2 = 0.25 * x * x;
-    x4 = x2 * x2;
-
-    if (fabs(x) < 10.0) {
-        *ber = 1.0;
-        r = 1.0;
-        for (m = 1; m <= 60; m++) {
-            r = -0.25 * r / (m * m) / pow((2.0 * m - 1.0), 2) * x4;
-            *ber += r;
-            if (fabs(r) < fabs(*ber)*eps) { break; }
-        }
-
-        *bei = x2;
-        r = x2;
-        for (m = 1; m <= 60; m++) {
-            r = -0.25 * r / (m * m) / pow((2.0 * m + 1.0), 2) * x4;
-            *bei += r;
-            if (fabs(r) < fabs(*bei)*eps) { break; }
-        }
-
-        *ger = -(log(x / 2.0) + el) * (*ber) + 0.25*pi*(*bei);
-
-        r = 1.0;
-        gs = 0.0;
-        for (m = 1; m <= 60; m++) {
-            r = -0.25 * r / (m * m) / pow((2.0 * m - 1.0), 2) * x4;
-            gs = gs + 1.0 / (2.0 * m - 1.0) + 1.0 / (2.0 * m);
-            *ger += r * gs;
-            if (fabs(r * gs) < fabs(*ger)*eps) { break; }
-        }
-
-        *gei = x2 - (log(x / 2.0) + el) * (*bei) - 0.25 * pi * (*ber);
-
-        r = x2;
-        gs = 1.0;
-        for (m = 1; m <= 60; m++) {
-            r = -0.25 * r / (m * m) / pow((2.0 * m + 1.0), 2) * x4;
-            gs = gs + 1.0 / (2.0 * m) + 1.0 / (2.0 * m + 1.0);
-            *gei += r * gs;
-            if (fabs(r * gs) < fabs(*gei)*eps) { break; }
-        }
-
-        *der = -0.25 * x * x2;
-        r = *der;
-        for (m = 1; m <= 60; m++) {
-            r = -0.25 * r / m / (m + 1.0) / pow((2.0 * m + 1.0), 2) * x4;
-            *der += r;
-            if (fabs(r) < fabs(*der)*eps) { break; }
-        }
-
-        *dei = 0.5 * x;
-        r = *dei;
-        for (m = 1; m <= 60; m++) {
-            r = -0.25 * r / (m * m) / (2.0 * m - 1.0) / (2.0 * m + 1.0) * x4;
-            *dei += r;
-            if (fabs(r) < fabs(*dei)*eps) { break; }
-        }
-
-        r = -0.25 * x * x2;
-        gs = 1.5;
-        *her = 1.5 * r - (*ber) / x - (log(x / 2.0) + el)*(*der) + 0.25*pi*(*dei);
-        for (m = 1; m <= 60; m++) {
-            r = -0.25 * r / m / (m + 1.0) / pow((2.0 * m + 1.0), 2) * x4;
-            gs = gs + 1.0 / (2.0 * m + 1.0) + 1.0 / (2 * m + 2.0);
-            *her += r * gs;
-            if (fabs(r * gs) < fabs(*her)*eps) { break; }
-        }
-
-        r = 0.5 * x;
-        gs = 1.0;
-        *hei = 0.5 * x - (*bei) / x - (log(x / 2.0) + el) * (*dei) - 0.25 * pi * (*der);
-        for (m = 1; m <= 60; m++) {
-            r = -0.25 * r / (m * m) / (2 * m - 1.0) / (2 * m + 1.0) * x4;
-            gs = gs + 1.0 / (2.0 * m) + 1.0 / (2 * m + 1.0);
-            *hei += r * gs;
-            if (fabs(r * gs) < fabs(*hei)*eps) { return; }
-        }
-    } else {
-        pp0 = 1.0;
-        pn0 = 1.0;
-        qp0 = 0.0;
-        qn0 = 0.0;
-        r0 = 1.0;
-        km = 18;
-        if (fabs(x) >= 40.0) km = 10;
-        fac = 1.0;
-        for (k = 1; k <= km; k++) {
-            fac = -fac;
-            xt = 0.25 * k * pi - trunc(0.125 * k) * 2.0 * pi;
-            cs = cos(xt);
-            ss = sin(xt);
-            r0 = 0.125 * r0 * pow((2.0 * k - 1.0), 2) / k / x;
-            rc = r0 * cs;
-            rs = r0 * ss;
-            pp0 += rc;
-            pn0 += fac * rc;
-            qp0 += rs;
-            qn0 += fac * rs;
-        }
-
-        xd = x / sqrt(2.0);
-        xe1 = exp(xd);
-        xe2 = exp(-xd);
-        xc1 = 1.0 / sqrt(2.0 * pi * x);
-        xc2 = sqrt(0.5 * pi / x);
-        cp0 = cos(xd + 0.125 * pi);
-        cn0 = cos(xd - 0.125 * pi);
-        sp0 = sin(xd + 0.125 * pi);
-        sn0 = sin(xd - 0.125 * pi);
-
-        *ger = xc2 * xe2 * (pn0 * cp0 - qn0 * sp0);
-        *gei = xc2 * xe2 * (-pn0 * sp0 - qn0 * cp0);
-        *ber = xc1 * xe1 * (pp0 * cn0 + qp0 * sn0) - (*gei) / pi;
-        *bei = xc1 * xe1 * (pp0 * sn0 - qp0 * cn0) + (*ger) / pi;
-
-        pp1 = 1.0;
-        pn1 = 1.0;
-        qp1 = 0.0;
-        qn1 = 0.0;
-        r1 = 1.0;
-        fac = 1.0;
-        for (int k = 1; k <= km; k++) {
-            fac = -fac;
-            xt = 0.25*k*pi - (int)(0.125 * k)*2.0*pi;
-            cs = cos(xt);
-            ss = sin(xt);
-            r1 = 0.125*r1*(4.0 - pow(2.0*k - 1.0, 2))/(k*x);
-            rc = r1*cs;
-            rs = r1*ss;
-            pp1 += fac*rc;
-            pn1 += rc;
-            qp1 += fac*rs;
-            qn1 += rs;
-        }
-        *her = xc2 * xe2 * (-pn1*cn0 + qn1*sn0);
-        *hei = xc2 * xe2 * (pn1*sn0 + qn1*cn0);
-        *der = xc1 * xe1 * (pp1*cp0 + qp1*sp0) - (*hei)/pi;
-        *dei = xc1 * xe1 * (pp1*sp0 - qp1*cp0) + (*her)/pi;
-    }
-    return;
-}
-
-
-inline void klvnzo(int nt, int kd, double *zo) {
-
-    // ====================================================
-    // Purpose: Compute the zeros of Kelvin functions
-    // Input :  NT  --- Total number of zeros
-    //          KD  --- Function code
-    //          KD=1 to 8 for ber x, bei x, ker x, kei x,
-    //                    ber'x, bei'x, ker'x and kei'x,
-    //                    respectively.
-    // Output:  ZO(M) --- the M-th zero of Kelvin function
-    //                    for code KD
-    // Routine called:
-    //          KLVNA for computing Kelvin functions and
-    //          their derivatives
-    // ====================================================
-
-    double ber, bei, ger, gei, der, dei, her, hei;
-    double rt0[9] = {0.0, 2.84891, 5.02622, 1.71854, 3.91467,
-                          6.03871, 3.77268, 2.66584, 4.93181};
-    double rt = rt0[kd];
-
-    for (int m = 1; m <= nt; m++) {
-        while (1) {
-            klvna(rt, &ber, &bei, &ger, &gei, &der, &dei, &her, &hei);
-            if (kd == 1) {
-                rt -= ber / der;
-            } else if (kd == 2) {
-                rt -= bei / dei;
-            } else if (kd == 3) {
-                rt -= ger / her;
-            } else if (kd == 4) {
-                rt -= gei / hei;
-            } else if (kd == 5) {
-                rt -= der / (-bei - der / rt);
-            } else if (kd == 6) {
-                rt -= dei / (ber - dei / rt);
-            } else if (kd == 7) {
-                rt -= her / (-gei - her / rt);
-            } else {
-                rt -= hei / (ger - hei / rt);
-            }
-
-            if (fabs(rt - rt0[kd]) <= 5e-10) {
-                break;
-            } else {
-                rt0[kd] = rt;
-            }
-        }
-        zo[m - 1] = rt;
-        rt += 4.44;
-    }
 }
 
 
@@ -6383,364 +5201,6 @@ void mtu12(int kf, int kc, int m, T q, T x, T *f1r, T *d1r, T *f2r, T *d2r) {
 }
 
 
-template <typename T>
-void pbdv(T x, T v, T *dv, T *dp, T *pdf, T *pdd) {
-
-    // ====================================================
-    // Purpose: Compute parabolic cylinder functions Dv(x)
-    //          and their derivatives
-    // Input:   x --- Argument of Dv(x)
-    //          v --- Order of Dv(x)
-    // Output:  DV(na) --- Dn+v0(x)
-    //          DP(na) --- Dn+v0'(x)
-    //          ( na = |n|, v0 = v-n, |v0| < 1,
-    //            n = 0,±1,±2,… )
-    //          PDF --- Dv(x)
-    //          PDD --- Dv'(x)
-    // Routines called:
-    //       (1) DVSA for computing Dv(x) for small |x|
-    //       (2) DVLA for computing Dv(x) for large |x|
-    // ====================================================
-
-    int ja, k, l, m, nk, nv, na;
-    T xa, vh, ep, f, f0, f1, v0, v1, v2, pd, pd0, pd1, s0;
-
-    xa = fabs(x);
-    vh = v;
-    v += copysign(1.0, v);
-    nv = (int)v;
-    v0 = v - nv;
-    na = abs(nv);
-    ep = exp(-0.25*x*x);
-    ja = 0;
-    if (na >= 1) { ja = 1; }
-    if (v >= 0.0) {
-        if (v0 == 0.0) {
-            pd0 = ep;
-            pd1 = x*ep;
-        } else {
-            for (l = 0; l <= ja; l++) {
-                v1 = v0 + l;
-                if (xa <= 5.8) {
-                    pd1 = dvsa(x, v1);
-                } else {
-                    pd1 = dvla(x, v1);
-                }
-                if (l == 0) {
-                    pd0 = pd1;
-                }
-            }
-        }
-        dv[0] = pd0;
-        dv[1] = pd1;
-        for (k = 2; k <= na; k++) {
-            *pdf = x*pd1 - (k+v0-1.0)*pd0;
-            dv[k] = *pdf;
-            pd0 = pd1;
-            pd1 = *pdf;
-        }
-    } else {
-        if (x <= 0.0) {
-            if (xa <= 5.8)
-            {
-                pd0 = dvsa(x, v0);
-                v1 = v0 - 1.0;
-                pd1 = dvsa(x, v1);
-            } else {
-                pd0 = dvla(x, v0);
-                v1 = v0 - 1.0;
-                pd1 = dvla(x, v1);
-            }
-            dv[0] = pd0;
-            dv[1] = pd1;
-            for (k = 2; k <= na; k++) {
-                pd = (-x*pd1+pd0)/(k-1.0-v0);
-                dv[k] = pd;
-                pd0 = pd1;
-                pd1 = pd;
-            }
-        } else if (x <= 2.0) {
-            v2 = nv + v0;
-            if (nv == 0) { v2 -= 1.0; }
-            nk = (int)(-v2);
-            f1 = dvsa(x, v2);
-            v1 = v2 + 1.0;
-            f0 = dvsa(x, v1);
-            dv[nk] = f1;
-            dv[nk-1] = f0;
-            for (k = nk-2; k >= 0; k--) {
-                f = x*f0 + (k-v0+1.0)*f1;
-                dv[k] = f;
-                f1 = f0;
-                f0 = f;
-            }
-        } else {
-            if (xa <= 5.8)
-            {
-                pd0 = dvsa(x, v0);
-            } else {
-                pd0 = dvla(x, v0);
-            }
-            dv[0] = pd0;
-            m = 100 + na;
-            f1 = 0.0;
-            f0 = 1e-30;
-            f = 0.0;
-            for (k = m; k >= 0; k--) {
-                f = x*f0 + (k-v0+1.0)*f1;
-                if (k <= na) { dv[k] = f; }
-                f1 = f0;
-                f0 = f;
-            }
-            s0 = pd0/f;
-            for (k = 0; k <= na; k++) { dv[k] *= s0; }
-        }
-    }
-    for (k = 0; k < na; k++) {
-        v1 = fabs(v0) + k;
-        if (v >= 0.0) {
-            dp[k] = 0.5*x*dv[k] - dv[k+1];
-        } else {
-            dp[k] = -0.5*x*dv[k] - v1*dv[k+1];
-        }
-    }
-    *pdf = dv[na-1];
-    *pdd = dp[na-1];
-    v = vh;
-    return;
-}
-
-
-template <typename T>
-inline void pbvv(T x, T v, T *vv, T *vp, T *pvf, T *pvd) {
-
-    // ===================================================
-    // Purpose: Compute parabolic cylinder functions Vv(x)
-    //          and their derivatives
-    // Input:   x --- Argument of Vv(x)
-    //          v --- Order of Vv(x)
-    // Output:  VV(na) --- Vv(x)
-    //          VP(na) --- Vv'(x)
-    //          ( na = |n|, v = n+v0, |v0| < 1
-    //            n = 0,±1,±2,… )
-    //          PVF --- Vv(x)
-    //          PVD --- Vv'(x)
-    // Routines called:
-    //       (1) VVSA for computing Vv(x) for small |x|
-    //       (2) VVLA for computing Vv(x) for large |x|
-    // ===================================================
-
-    int ja, k, kv, l, m, na, nv;
-    T f, f0, f1, pv0, q2p, qe, s0, v0, v1, v2, vh, xa;
-
-    const T pi = 3.141592653589793;
-
-    xa = fabs(x);
-    vh = v;
-    v += copysign(1.0, v);
-    nv = (int)v;
-    v0 = v - nv;
-    na = abs(nv);
-    qe = exp(0.25*x*x);
-    q2p = sqrt(2.0/pi);
-    ja = 0;
-    if (na >= 1) { ja = 1; }
-    f = 0.0;
-    if (v <= 0.0) {
-        if (v0 == 0.0) {
-            if (xa <= 7.5) {
-                pv0 = vvsa(x, v0);
-            } else {
-                pv0 = vvla(x, v0);
-            }
-            f0 = q2p*qe;
-            f1 = x*f0;
-            vv[0] = pv0;
-            vv[1] = f0;
-            vv[2] = f1;
-        } else {
-            for (l = 0; l <= ja; l++) {
-                v1 = v0-l;
-                if (xa <= 7.5) {
-                    f1 = vvsa(x, v1);
-                } else {
-                    f1 = vvla(x, v1);
-                }
-                if (l == 0) { f0 = f1; }
-            }
-            vv[0] = f0;
-            vv[1] = f1;
-        }
-        kv = 2;
-        if (v0 == 0.0) { kv=3; }
-        for (k = kv; k <= na; k++) {
-            f= x*f1 + (k-v0-2.0)*f0;
-            vv[k] = f;
-            f0 = f1;
-            f1 = f;
-        }
-    } else {
-        if ((x >= 0.0) && (x <= 7.5)) {
-            v2 = v;
-            if (v2 < 1.0) { v2 = v2+1.0; }
-            f1 = vvsa(x, v2);
-            v1 = v2 - 1.0;
-            kv = (int)v2;
-            f0 = vvsa(x, v1);
-            vv[kv] = f1;
-            vv[kv - 1] = f0;
-            for (k = kv-2; k >= 0; k--) {
-                f = x*f0 - (k+v0+2.0)*f1;
-                if (k <= na) { vv[k] = f; }
-                f1 = f0;
-                f0 = f;
-            }
-        } else if (x > 7.5) {
-            pv0 = vvla(x, v0);
-            m = 100 + abs(na);
-            vv[1] = pv0;
-            f1 = 0.0;
-            f0 = 1.0e-40;
-            for (k = m; k >= 0; k--) {
-                f= x*f0 - (k+v0+2.0)*f1;
-                if (k <= na) { vv[k] = f; }
-                f1 = f0;
-                f0 = f;
-            }
-            s0 = pv0/f;
-            for (k = 0; k <= na; k++) {
-                vv[k] *= s0;
-            }
-        } else {
-            if (xa <= 7.5) {
-                f0 = vvsa(x, v0);
-                v1 = v0 + 1.0;
-                f1 = vvsa(x, v1);
-            } else {
-                f0 = vvla(x, v0);
-                v1 = v0 + 1.0;
-                f1 = vvla(x, v1);
-            }
-            vv[0] = f0;
-            vv[1] = f1;
-            for (k = 2; k <= na; k++) {
-                f = (x*f1 - f0)/(k + v0);
-                vv[k] = f;
-                f0 = f1;
-                f1 = f;
-            }
-        }
-    }
-    for (k = 0; k < na; k++) {
-        v1 = v0 + k;
-        if (v >= 0.0) {
-            vp[k] = 0.5*x*vv[k] - (v1+1.0)*vv[k+1];
-        } else {
-            vp[k] = -0.5*x*vv[k] + vv[k+1];
-        }
-    }
-    *pvf = vv[na-1];
-    *pvd = vp[na-1];
-    v = vh;
-    return;
-}
-
-
-template <typename T>
-void pbwa(T a, T x, T *w1f, T *w1d, T *w2f, T *w2d) {
-
-    // ======================================================
-    // Purpose: Compute parabolic cylinder functions W(a,±x)
-    //          and their derivatives
-    // Input  : a --- Parameter  ( 0 ≤ |a| ≤ 5 )
-    //          x --- Argument of W(a,±x)  ( 0 ≤ |x| ≤ 5 )
-    // Output : W1F --- W(a,x)
-    //          W1D --- W'(a,x)
-    //          W2F --- W(a,-x)
-    //          W2D --- W'(a,-x)
-    // Routine called:
-    //         CGAMA for computing complex gamma function
-    // ======================================================
-
-    int k, L1, L2;
-    T d[80], d1, d2, dl, f1, f2, g1, g2, h[100], h0, h1, hl, r, r1,\
-           y1d, y2d, y1f, y2f;
-    std::complex<T> ug, vg;
-    const T eps = 1e-15;
-    const T p0 = 0.59460355750136;
-
-    if (a == 0.0) {
-        g1 = 3.625609908222;
-        g2 = 1.225416702465;
-    } else {
-        ug = cgama(std::complex<T>(0.25, 0.5*a), 1);
-        g1 = std::abs(ug);
-        vg = cgama(std::complex<T>(0.75, 0.5*a), 1);
-        g2 = std::abs(vg);
-    }
-    f1 = sqrt(g1/g2);
-    f2 = sqrt(2.0*g2/g1);
-    h0 = 1.0;
-    h1 = a;
-    h[0] = a;
-    for (L1 = 2; L1 <= 100; L1++) {
-        hl = a*h1 - 0.25*(2*L1 - 2.0)*(2*L1 - 3.0)*h0;
-        h[L1 - 1] = hl;
-        h0 = h1;
-        h1 = hl;
-    }
-    y1f = 1.0;
-    r = 1.0;
-    for (k = 1; k <= 100; k++) {
-        r = 0.5*r*x*x/(k*(2.0*k - 1.0));
-        r1 = h[k - 1]*r;
-        y1f += r1;
-        if ((fabs(r1) <= eps*fabs(y1f)) && (k > 30)) { break; }
-    }
-    y1d = a;
-    r = 1.0;
-    for (k = 1; k < 100; k++) {
-        r = 0.5*r*x*x/(k*(2.0*k + 1.0));
-        r1 = h[k]*r;
-        y1d += r1;
-        if ((fabs(r1) <= eps*fabs(y1d)) && (k > 30)) { break; }
-    }
-    y1d *= x;
-    d1 = 1.0;
-    d2 = a;
-    d[0] = 1.0;
-    d[1] = a;
-    for (L2 = 3; L2 <= 80; L2++) {
-        dl = a*d2 - 0.25*((2*L2-1) - 2.0)*((2*L2-1) - 3.0)*d1;
-        d[L2 - 1] = dl;
-        d1 = d2;
-        d2 = dl;
-    }
-    y2f = 1.0;
-    r = 1.0;
-    for (k = 1; k < 80; k++) {
-        r = 0.5*r*x*x/(k*(2.0*k + 1.0));
-        r1 = d[k]*r;
-        y2f += r1;
-        if ((fabs(r1) <= eps*fabs(y2f)) && (k > 30)) { break; }
-    }
-    y2f *= x;
-    y2d = 1.0;
-    r = 1.0;
-    for (k = 1; k < 80; k++) {
-        r = 0.5*r*x*x/(k*(2.0*k - 1.0));
-        r1 = d[k]*r;
-        y2d += r1;
-        if ((fabs(r1) <= eps*fabs(y2d)) && (k > 30)) { break; }
-    }
-    *w1f = p0*(f1*y1f - f2*y2f);
-    *w2f = p0*(f1*y1f + f2*y2f);
-    *w1d = p0*(f1*y1d - f2*y2d);
-    *w2d = p0*(f1*y1d + f2*y2d);
-    return;
-}
-
-
 inline double psi_spec(double x) {
 
     // ======================================
@@ -7968,99 +6428,6 @@ inline void sphy(T x, int n, int *nm, T *sy, T *dy) {
         dy[k] = sy[k - 1] - (k + 1.0) * sy[k] / x;
     }
     return;
-}
-
-
-inline double vvla(double x, double va) {
-
-    // ===================================================
-    // Purpose: Compute parabolic cylinder function Vv(x)
-    //          for large argument
-    // Input:   x  --- Argument
-    //          va --- Order
-    // Output:  PV --- Vv(x)
-    // Routines called:
-    //       (1) DVLA for computing Dv(x) for large |x|
-    //       (2) GAMMA2 for computing Г(x)
-    // ===================================================
-
-    int k;
-    double pv, qe, a0, r, x1, gl, dsl, pdl;
-
-    const double pi = 3.141592653589793;
-    const double eps = 1e-12;
-
-    qe = exp(0.25*x*x);
-    a0 = pow(fabs(x), -va-1)*sqrt(2.0/pi)*qe;
-    r = 1.0;
-    pv = 1.0;
-    for (k = 1; k <= 18; k++) {
-        r = 0.5 * r * (2.0*k + va - 1.0)*(2.0*k + va)/(k*x*x);
-        pv += r;
-        if (fabs(r/pv) < eps) { break; }
-    }
-    pv *= a0;
-    if (x < 0.0) {
-        x1 = -x;
-        pdl = dvla(x1, va);
-        gl = gamma2(-va);
-        dsl = sin(pi*va)*sin(pi*va);
-        pv = dsl*gl/pi*pdl - cos(pi*va)*pv;
-    }
-    return pv;
-}
-
-
-inline double vvsa(double x, double va) {
-
-    // ===================================================
-    // Purpose: Compute parabolic cylinder function Vv(x)
-    //          for small argument
-    // Input:   x  --- Argument
-    //          va --- Order
-    // Output:  PV --- Vv(x)
-    // Routine called : GAMMA2 for computing Г(x)
-    // ===================================================
-
-    double a0, fac, g1, ga0, gm, gw, r, r1, sq2, sv, sv0, v1, vb0, vm, pv;
-
-    const double eps = 1.0e-15;
-    const double pi = 3.141592653589793;
-    const double ep = exp(-0.25 * x * x);
-    double va0 = 1.0 + 0.5 * va;
-
-    if (x == 0.0) {
-        if (((va0 <= 0.0) && (va0 == (int)va0)) || va == 0.0) {
-            pv = 0.0;
-        } else {
-            vb0 = -0.5 * va;
-            sv0 = sin(va0 * pi);
-            ga0 = gamma2(va0);
-            pv = pow(2.0, vb0) * sv0 / ga0;
-        }
-    } else {
-        sq2 = sqrt(2.0);
-        a0 = pow(2.0, -0.5 * va) * ep / (2.0 * pi);
-        sv = sin(-(va + 0.5) * pi);
-        v1 = -0.5 * va;
-        g1 = gamma2(v1);
-        pv = (sv + 1.0) * g1;
-        r = 1.0;
-        fac = 1.0;
-
-        for (int m = 1; m <= 250; m++) {
-            vm = 0.5 * (m - va);
-            gm = gamma2(vm);
-            r = r * sq2 * x / m;
-            fac = -fac;
-            gw = fac * sv + 1.0;
-            r1 = gw * r * gm;
-            pv += r1;
-            if ((fabs(r1 / pv) < eps) && (gw != 0.0)) { break; }
-        }
-        pv *= a0;
-    }
-    return pv;
 }
 
 }
